refactor - chanve valor from double to Bigdecimal

+++ src/main/java/pt/whd/readFile/Pedido.kt
@@ -1,3 +1,5 @@
 package pt.whd.readFile
 
-class Pedido(val id: Long, val estado: String, val valor: Double)
+import java.math.BigDecimal
+
+class Pedido(val id: Long, val estado: String, val valor: BigDecimal)
